correct double

+++ lab2.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <math.h>
 
-void qwrt(int a, int b, int c)
+void qwrt(double a, double b, double c)
 {
-	int d;
+	double d;
 	double x1, x2;
 	d = b * b - 4 * a * c;
 	if (d < 0)
@@ -13,21 +13,21 @@ void qwrt(int a, int b, int c)
 	else if (d == 0)
 	{
 		x1 = -b / (2 * a);
-		printf("\nRoot: %f\n", x1);
+		printf("\nRoot: %Lf\n", x1);
 	}
 	else if (d > 0)
 	{
 		x1 = (-b + sqrt(d)) / (2 * a);
 		x2 = (-b - sqrt(d)) / (2 * a);
-		printf("\nRoots: %f %f\n", x1, x2);
+		printf("\nRoots: %Lf %Lf\n", x1, x2);
 	}
 }
 
 int main()
 {
-	int a, b, c;
+	double a, b, c;
 	printf("Enter a, b, c: \n");
-	scanf("%d%d%d", &a, &b, &c);
+	scanf("%Lf%Lf%Lf", &a, &b, &c);
 	qwrt(a, b, c);
 	return 0;
 }
